Moved git tests to test.py
